+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 These are the changes to each version that has been released
 on the official Visual Studio extension gallery.
 
+## 2.1.1
+ - [x] Added events to definition file, some IntelliSense tuning
+
+## 2.1
+ - [x] Support for globally registered component definition file
+
 ## 2.0
  - [x] Updated autocompletion to VueJS 2.6